AGC/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

torch==2.10
digeo
robust-laplacian
pyyaml
matplotlib
polyscope
open3d
trimesh
tqdm
